Add verification of signature when adding to tx (#174)
Add code to verify a signature when added to a transaction.

This is valuable for scenarios where a party receives a signature from another party and wants to add this signature to a transaction. The first party needs to make sure that the signature from the other party is valid.

This will become apparent once the transaction is submitted to horizon – it will fail if the signature is invalid. However, for smart contract schemes (say, payment channels) this will usually happen much later.