Test equality directly for the XDR switch value.
For some reason, the build breaks on Windows 8.1 unless you test
the switch value directly. Otherwise, the === comparison on the
switch() fails, and thus the account isn't encoded correctly.

We'd expect the actual unit test line to fail as well because of
this (i.e. any place we do
`expect(xdr.switch()).to.equal(xxx.switch())`) but it seems like
the chai framework's expect.equal is different than ===. This is
a mystery, and I imagine testing on Windows 10 would solve this
problem, but that's another battle for another day.