Replace expect.eql (which is a deep-equal) with Buffer.equals:
This is yet another instance of the CI environment behaving
strangely: you'd expect deep-equal to, well, deep-equal, but w/o
this change we see errors like this:

  AssertionError: expected
  <Buffer 3f 0c 34 bf 93 ad 0d 99 71 d0 4c cc> to deeply equal
  <Buffer 3f 0c 34 bf 93 ad 0d 99 71 d0 4c cc>

which is obviously nonsense. However, using the native
Buffer.equals method resolves this issue and lets the CI pass.

Again, this could probably be fixed by bumping the minimum
environment expectations to Windows 10+, but some are resistant to
this new minimum.