Modifies tests to expect that we prevent forcing muxed XDRs.
Currently, we were creating an `xdr.MuxedAccount(pubkey=G..., ID=0)` object
(where the `G...` is the XDR-encoded public key bytes not the string, to be
clear) if you called `decodeAddressToMuxedAccount(G..., true)`. In other words,
we would always force G... to be fully-muxed w/ ID=0 if users opted in to
muxing.

This behavior is broken for scenarios in which users mix muxed and unmuxed
parameters. For example, creating a payment operation with a muxed destination
and an unmuxed source account will result in XDR containing a muxed source
account with an ID of zero, because they passed `withMuxing: true`.

This isn't faithful to what the API user expected, since the XDR would be
identical in the case where they made a payment with muxed destination *and*
source with id=0 set explicitly.

The new test `supports mixing muxed and unmuxed properties` will fail without
the necessary changes, and the `decodes to an empty muxed account` is no longer
expected behavior, so it now ensures `lets G... accounts pass through`.