Drop tests for 17.x because of incompatibility:
Node 17.x upgrades OpenSSL to v3 which basically changes some stuff
related to hashing in a way that isn't backwards compatible.

Relevant link:
  https://github.com/webpack/webpack/issues/14532#issuecomment-947012063

This (a) causes breakages that aren't relevant to us and (b) has
a dangerous workaround:

  export NODE_OPTIONS=--openssl-legacy-provider

If we do this workaround here, we are basically brushing the issue
under the rug. The longer term solution is basically to let the
ecosystem deal with it for now and when we need to make 18.x the
minimum, to deal with it in turn.

Since our minimum is 14.x for now, there's no reason to do
17.x-specific upgrades or workarounds.